rainforest: extract the crc32 calculation code into rf_crc32
This one is huge and makes the code less readable. Let's move it into
its own file which is included as-is at the same place. It was not
changed at all.
